backend/requirements.txt
bcrypt==4.3.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
# numpy==2.0.0
openpyxl==3.1.5 
# pandas==2.2.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
cryptography==41.0.3
requests==2.31.0